Resolve issues when using get_ready()/handled_ready() construct
We used to just removing the first item from that list, thus causing issues if one would call conn_setevent()
in the meantime to 'shut up' that connection. This is used for letting that socket wait, not processing it further
till we start caring for it again, eg with jumpbox when we finally receive a /pull/ or are still otherwise processing
the request.