Restructure request parser to make argument parsing more efficient
This new variant does go through the whole URL but does not error out on problems (it does log them).
We also do not keep the unused arguments in argsplit as we know which ones we are looking for
This avoids having to need to maintain a pointer of all arguments, even if we don't care for them.